+++ src/index.ts
@@ -1,4 +1,9 @@
-import { vec2, Vector2 } from './vector2';
+import { vec, Vector } from './vector';
+import { vtx, Vertex } from './vertex';
 import { edge, Edge } from './edge';
 
-export { vec2, Vector2, edge, Edge };
+export { 
+    vec, Vector, 
+    vtx, Vertex, 
+    edge, Edge 
+};